+++ core/designsystem/src/main/java/co/kr/tnt/designsystem/component/Dialog.kt
@@ -17,6 +17,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
 import androidx.compose.ui.window.DialogProperties
+import co.kr.tnt.designsystem.component.button.TnTTextButton
 import co.kr.tnt.designsystem.theme.TnTTheme
 import co.kr.tnt.designsystem.util.ButtonSize
 import co.kr.tnt.designsystem.util.ButtonType
